ForbiddenNegativeBitshift: fix false positive and underreporting
### Fix underreporting

The `ForbiddenNegativeBitshift` was only listening to "bitshift right", while the change made in PHP 7.0 also affects:
* bitshift left
* bitshift left assignments
* bitshift right assignments

These tokens have now been added to the array to listen to and negative bitshifting for these will now also be reported.

### Fix false positives

The sniff should err on the side of caution and only report when it can be determined with some form of certainty that the operand is a negative number.

I.e. calculations with variables which may or may not result in a negative number should not be reported.

To this end, the new `isNegativeNumber()` utility method is used to better determine whether the operand to the right of the operator is a negative number.

### Other

Includes a minor improvement to the error message, which will now actually show the right operand as used to determine that the bitshift was done by a negative number.

Includes additional unit tests.

Refs:
* http://php.net/manual/en/migration70.incompatible.php#migration70.incompatible.integers.negative-bitshift
* https://wiki.php.net/rfc/integer_semantics

Fixes 294
Fixes 446