Add a WordPress specific PHPCompatibility ruleset
Frameworks/CMSes often provide backfills for PHP native functionality which is not available in the minimum PHP version supported by the framework/CMS.

As suggested in 530, this is a first PR adding a framework specific ruleset to PHPCompatibility.

To give some context:
In the WordPress community, there is currently a project under way called [Tide](https://make.wordpress.org/tide/).
This project uses the PHPCompatibility library to analyse all the plugins (50.000+) and themes (3.000+) hosted on wordpress.org and will in the near future be used to start displaying PHP compatibility information for each of them on the wordpress.org website.

With that in mind, I expect a significant number of developers in the WordPress community will want to start using PHPCompatibility as part of their QA process.

WordPress provides a number of backfills. Now if each of these developers needs to figure out the backfills in WP themselves, I expect we'll see a significant number of support requests related to this while our time is better spend on writing/improving sniffs.

So providing a ready to use framework specific ruleset will make life easier both for the WP devs who want to start using PHPCompatibility as well as for ourselves.

Notes:
* Adds a new dedicated subdirectory `framework-rulesets` with initially a `wordpress.xml` ruleset.
* Adds a new section to the `README` which explains how projects can use framework/CMS specific rulesets as hosted in this repository.
* Adds a new section to the `CONTRIBUTING` document with some basic guidelines for adding/contributing to these rulesets.

Related to 530. I suggest leaving the issue open for now until at least a second framework/CMS specific ruleset has been added.

-----

Additional considerations:

As the path to these custom rulesets will - depending on the chosen installation method - not always be predictable, using these custom rulesets will not be so easy for non-Composer users of PHPCompatibility.
And even from Composer users, I expect we'll see some support requests.

The alternative would have been to add a `PHPCompatibility-WP` directory with a `ruleset.xml` file which would allow the ruleset to be registered with PHPCS under its own name and used as `--standard=PHPCompatibility-WP`.

The reasons I did not do so at this time, is that - if a lot of different framework specific ruleset are added -, this would mean that the PHPCompatibility repository would add a lot of named rulesets to a PHPCS install making the list of installed standards huge.

If/when more framework/CMS specific rulesets are added and they prove popular, I would suggest to consider moving the PHPCompatibility repository to a dedicated organisation.
This will then allow for more repositories to be added to the organisation, opening the way for `PHPCompatibility-WP`, `PHPCompatibility-Drupal`, `PHPCompatibility-Laraval` (etc) repositories which *would* add named rulesets for their respective frameworks.