Composer: allow for PHPUnit 7.x
PHPUnit 7.x has come out in the mean time and is already in use on some of the Travis images.

Looks like the current unit test suite will work on PHPUnit 7.x without issues, so let's indicate that.

Also makes sure that if PHPUnit 4.x is used, it will get a compatible version as per 511.