NewHeredocInitialize: extend the NewScalarExpressions sniff
The logic within the original `NewHeredocInitialize` sniff did not allow for multi-declarations nor for examining the default values of function parameters.

While not (well) documented, these features are supported and do allow initialization with heredocs since PHP 5.3.

As the logic to determine whether a token needs to be examined is the same for this sniff and for the new `NewScalarExpressions` sniff, letting the sniff extend the `NewScalarExpressions` sniff prevents code duplication.