NewKeywords: fix too aggressive accounting for PHPCS cross-compat
A number of tokens are not correctly back-filled across PHP versions by PHPCS.
To that end, a _token translation table_ is used to recognize tokens tokenized as `T_STRING` as the keyword anyway.

At the same time, PHPCS retokenizes sometimes too agressively.

Having said that, most PHP keywords can be safely used as variable (and property) names and as of PHP 7, as the name of class constants.
However, these are tokenized as `T_STRING` which leads to false positives from this sniff.

At the same time, keywords in PHP are case-insensitive, so the cross-compat layer should allow for mixed case / uppercase use of the (non magic-constant) keywords as well.

This commit fixes both.
Includes unit tests.