:sparkles: PHP 7.3: new sniff to detect trailing comma's in function calls, isset() and unset()
Trailing comma's in function calls, method calls, `isset()` and `unset()` will be allowed as of PHP 7.3.

Refs:
* https://wiki.php.net/rfc/trailing-comma-function-calls
* https://github.com/php/php-src/commit/b591c329ee3129adbdc35141bb1542d119f7a2a1

I've kept the name of the sniff relatively generic to allow for adding more constructs in the future in case trailing comma's will start to be allowed in, for instance, `list()`.

Includes unit tests which are largely based on the example code in the RFC/unit tests in PHP core.