New AbstractFunctionCallParameterSniff
This adds a new abstract base class to do the initial checks necessary of whether a T_STRING token is a function/method call and whether the function called is one of the target functions.

Child classes are expected to set:
* The `$targetFunctions` array to indicate the names of the functions (as array keys) which the child class is looking for.
* [Optional] the `$isMethod` property to `true` if they are looking for method calls.

Child classes are expected to implement a `processParameters()` method to examine the contents of the parameters passed to the function call.