:sparkles: PHP 5.6: New sniff to detect passing deprecated $type values to iconv_get_encoding
"The iconv and mbstring configuration options related to encoding have been deprecated in favour of default_charset."

Refs:
* http://php.net/manual/en/migration56.deprecated.php#migration56.deprecated.iconv-mbstring-encoding
* https://wiki.php.net/rfc/default_encoding

Note:
As it is unclear which MBString functions to target, the sniff only checks the affected `iconv` function.
Based on the commit and the other documentation, I get the impression that for MBString, this only affects the ini directives.

Fixes 475