:sparkles: PHP 7.0: New sniff to detect list()s affected by the assignment order change
>  `list()` no longer assigns variables in reverse order
>
> `list()` will now assign values to variables in the order they are defined, rather than reverse order. In general, this only affects the case where `list()` is being used in conjunction with the array [] operator

Refs:
* http://php.net/manual/en/migration70.incompatible.php#migration70.incompatible.variable-handling.list.order
* https://wiki.php.net/rfc/abstract_syntax_tree#changes_to_list
* http://php.net/manual/en/function.list.php

Note: in contrast to the changelog entry, this actually affects all lists which use the same variable twice in the `list()` construct, so that is the premise the sniff has been based on now.
See: https://3v4l.org/LpUl6

Open question:
I've set the sniff up now to throw an error if non-unique list variables are found and PHP 7.0 or higher is supported.
It could be argued that this sniff should always throw an error when non-unique list variables are encountered as this will make the resulting variable values unpredictable.
It could also be argued that the sniff should only throw an error if the `testVersion` contains both PHP 5.x as well as 7.x. I.e. no error when only PHP 5 is supported and no error when only PHP 7 is supported, but an error when 5.5-7.2 is supported.

Opinions ?