Sniff::getFunctionCallParameters(): correctly handle closures when passed as param
Closures can be declared within function calls when the function expects a call-back.
This situation was only partially handled within the `Sniff::getFunctionCallParameters()` method.

Most of the time when a comma is encountered within a closure, it will be at a different nesting level than the original function call/array, which meant it would be disregarded and the method would return the correct results.
However, there are some, albeit rare, situations in which the comma would be at the target nesting level, causing the method to split the closure into two or more "parameters".

I came across this while working on a PR for WPCS which uses the same utility function.
As far as I know, this bug has not caused any issues for any of the PHPCompatibility sniffs so far, but the situation should be handled correctly even so.

Includes unit tests.