New sniff to check minimum supported/recommended PHPCS version
As it has been decided that PHPCS 1.5.x support will be dropped in PHPCompatibility 9.0.0, this will give users of PHPCS versions below the next minimum recommended version - 2.6.0 - advance notice.
`2.6.0` is already mentioned in the readme as the minimum recommended version and has been for a while (since July 2017).
`2.3.0` will be the next minimum _supported_ version. A note to that effect has also been added to the Readme.

The sniff will throw a `warning` when the PHPCS version PHPCompatibility is run on, is below the minimum recommended version and an error when the PHPCS version is below the minimum supported version.

The error message contains delimiter lines to improve the readability of the (very long) message.
Unfortunately, using line breaks is not an option as the implementation of that in PHPCS itself (supported since 2.4.0) is too buggy - see /squizlabs/PHP_CodeSniffer 2093 for more details.

The error codes are set up to include the version numbers. This will make sure that if a user silences the message from within their custom ruleset and the minimum supported/recommended PHPCS version(s) for PHPCompatibility change, the notices will be thrown again.

This sniff is not unit tested as the unit test would need to contain the same logic as the sniff, so wouldn't actually test anything.

The PR also includes a new utility method to get access to the `reportWidth` and `showSources` config settings.
The existing `PHPCSHelper::getConfigData()` turns out to only have access to the info provided via `--config-set` and `--runtime-set`, not to info provided via other command-line/custom ruleset arguments.

Related to 688