Sniffs: remove work-arounds for old format of the PHPCS (token) arrays
In PHPCS 1.x the token arrays were in the following format:
```php
$array = array(
    T_TOKENNAME,
);
```
In PHPCS 2.0 and higher, the token arrays are in the following format:
```php
$array = array(
    T_TOKENNAME => T_TOKENNAME,
);
```

This means that work-around like `array_flip()`/`array_combine()` to enable using `isset()` are no longer necessary. It also means that some `in_array()`s can be replaced with `isset()`s.