NewGeneratorReturn: bug fix for generator return with nested functions
The sniff didn't take well enough into account that generator functions could be nested within other (anonymous) functions or could have other nested constructs within the generator within which a `return` statement was fine.

Previously, it would check if the `yield` had a (named) `function` scope and if not, if it had a anonymous function scope.
This caused the false positive as reported by mjrider as for a `yield` in a closure nested within a named function, the named function would be examined, not the closure.

The fix in this PR changes that logic to find the deepest nested function òr closure within which the `yield` exist.

Additionally, this fix makes sure that the sniff doesn't get confused over `return` statements found in even deeper nested scopes, such as a closure nested within the generator.

Includes additional unit tests.

Fixes 724