Utility Function Tests: change test case file loading principle
The unit test files testing the utility functions in the `Sniff` base class via the `CoreMethodTestFrame` class, all just use one unit test case file.
This unit test case file has been moved to be in the same directory and renamed to the same name as the unit test file, but using the `.inc` file extension.

This now makes the unit test case file names predictable, meaning that the file name no longer needs to be set in each unit test file, but can be determined based on the name of the unit test class currently active.

Includes:
* Adjusting the namespace and classname of the `CoreMethodTestFrame` class (after the file rename).
* Loading the `TestHelperPHPCompatibility` via a `use` statement rather than via a `require_once`.
* Adjusting the namespace of the `TestHelperPHPCompatibility` class (after the file rename).