Sniff Tests: change BaseSniffTest::sniffFile() method
... to deal with the new directory structure.

This changes the method signature of the `BaseSniffTest::sniffFile()` method.

Where previously, it was expected that `$filename` would be passed with a partial path relative to the directory containing the `BaseSniffTest.php` file, functions call are now expected to pass an absolute path to the test case file.

The changed method allows for test files to pass the absolute path to itself using `__FILE__` when the test file only uses one test case file.
In that case, the test case file is presumed to be named the same as the test file, but with an `.inc` file extension.

The caching of already scanned files will now use the absolute file path as the cache key. This will allow for sniffs in different categories to have the same name.
While that practice is still discouraged, it would no longer give incorrect results.