NewOperators: Update the version number for T_COALESCE_EQUAL
While this language construct is already tokenized correctly in PHPCS, PHP itself has not yet implemented it.

The RFC has been approved a while back and it was slated to go into PHP 7.2, but as of PHP 7.3 RC2 no implementation has been merged and it is not expected to go into PHP 7.3 anymore.

As PHPCS already recognized the token, it was already added to the `NewOperators` sniff a while back.

All this PR does is up the PHP version number used for the compares.

Also see 523.