QA: add `use` statements for the aliased PHPCS and PHPUnit classes
For those classes which have 1-on-1 aliases across PHPCS versions, we can use `use ... as ...` statements to import the aliased class and use a shorter reference inline.

This applies to:
* `PHP_CodeSniffer_File` - aliased as `File` which is the PHPCS 3.x class name.
* `PHP_CodeSniffer_Tokens` - aliased as `Tokens` which is the PHPCS 3.x class name.
* `PHP_CodeSniffer_Exception` - in PHPCS 3.x, this class is called `PHP_CodeSniffer\Exceptions\RuntimeException`, aliased to `PHPCS_Exception` to prevent confusion with the PHP native `RuntimeException` class.
* `Generic_Sniffs_NamingConventions_CamelCapsFunctionNameSniff` - aliased as `PHPCS_CamelCapsFunctionNameSniff`.
* `PHP_CodeSniffer_Standards_AbstractScopeSniff` - aliased as `PHPCS_AbstractScopeSniff` to prevent confusion with the PHPCompatibility native abstract sniffs.

Note: this is not possible for those classes which don't have a direct counter-part between PHPCS 2 and PHPCS 3!

Additionally, the references to the PHPUnit class - which is aliased within the `phpunit-bootstrap.php` file - have also been changed to `use` statements.