RemovedPHP4StyleConstructors: fix bug with nested constructs
Inspired by upstream issue 2178, I checked that the same bug didn't exist in the related PHPCompatibility sniff and found that it did.

Fixed by simply skipping to the function closer after examining each function declaration as the content of the function is not our concern.

This should also make the sniff faster.

Notes:
* In PHPCS 2.3.x, anonymous classes do not yet have their own token, so the class name determination would be incorrect as the upstream method searches for the first `T_STRING` and in this case, that would be the name of the first function in the class, nested or otherwise.
    Added a work-around for this.
* Interfaces can declare deprecated constructors as well and should trigger the error.

Includes unit test for both interfaces, classes with abstract methods as well as for the original case: functions in a nested anonymous class.

Also added some additional unit tests to raise code coverage.