:rocket: Build/Travis: test builds against PHP 7.3
Once PHP 7.3-beta came out, the `nightly` build on Travis became PHP 7.4-dev and the builds haven't been tested against PHP 7.3 for months now.

Luckily, Travis has, in the mean time, deemed it appropriate to set up a PHP 7.3 alias now RC3 is out, so I've added PHP 7.3 to the matrix.

**Important note**:
PHP 7.3+ is only fully supported in combination with PHPCS 2.9.2 and 3.3.1+.
When running PHPCompatibility against PHPCS < 2.9.2 or PHPCS 3.x < 3.3.1, PHP warnings about using `continue` within `switch` will be thrown, but will in most cases not affect the functioning of the sniffs.
While on PHPCS 2.x, the tests won't fail, on PHPCS 3.x < 3.3.1, they will, as the error handling during PHPCS runs has changed in PHPCS 3.x.
In other words, for PHP 7.3, we can safely run the unit tests on PHPCS 2.x and PHPCS >= 3.3.1, but PHPCS 3.0 - 3.3.0 can not be tested.

Includes updating to `README` with a note about PHP 7.3 support.

N.B.: It would have been nice to move the coverage builds onto 7.3 at this time as that should raise the coverage a little as the PHP 7.3 specific code in the `NewFlexibleHeredocNowdoc` sniff should then show as covered.
However, that is not yet possible at this time as there is _no code coverage driver available_ for PHP 7.3 in the PHPUnit version included on the PHP 7.3 Travis images at this time.
See https://travis-ci.org/jrfnl/PHPCompatibility/jobs/465544658#L808