`Sniff:getTestVersion()`: throw Exception instead of PHP native Warning
Ok, so this one will need some explaining and is actually about allowing the unit tests to pass on PHP 7.3.

First off, the underlying problem is that PHPCS isn't fully compliant with PHP 7.3 until version 3.3.1 - see upstream PR 2086.
PHPCS prior to version 3.3.1, has a couple of `continue` statements within `switch` control structures causing PHP 7.3 to throw a warning. One of these is in the code for the `Tokenizer` class which is used to enhance the token array for every file being analysed and therefore unavoidable.

In PHPCS 2.x, this doesn't cause too much of a problem as PHPCS handled warnings encountered during the tokenizer runs differently.
In PHPCS 3.x, however, it becomes a problem as when such a warning will get thrown during the unit test run, PHPUnit will mark the first test which throws a warning as `Error` and fail the build.
Skipping select tests for the PHP 7.3 / PHPCS 3.0 < 3.3.1 combi will not help as that just moves the error to the next unit test with the same net effect.

Now, PHPUnit has a `convertWarningsToExceptions` setting which defaults to `true`. Changing that setting to `false` will allow the unit test run to pass as in that case, the first test which sees the warning will be marked as `Risky` for having output, but will not fail the build.

The next problem we then run into, is that PHPCompatibility itself throws a warning in the `getTestVersion()` method.
With the `convertWarningsToExceptions` setting set to `false`, those warnings will be thrown and the unit tests for that method which relied on the warnings being converted to exceptions will now fail.

That, however, is fixable.

So, I've added a new `PHPCompatibility\Util\TestVersionException` class and turned the warnings which were being thrown from the `getTestVersion()` method into Exceptions.

This Exception unfortunately can't extend the PHPCS native `RuntimeException` as PHPCS native exceptions weren't added until PHPCS 3.

Having said that, the current implementation allows the `convertWarningsToExceptions` setting to be set to `false` and our unit tests to still pass.
For PHP 7.3 in combination with PHPCS 3.0 < 3.3.1, this means that the very first test **will** be marked as risky, but other than that, the builds should pass.

**Note**: there is one additional caveat. PHPCS 2.x has a bug in the error handling and will throw an `Undefined offset: 0` error now whenever we throw the new Exception in PHPCS 2.x.
That warning will _only_ be thrown when someone sets an invalid `testVersion` though, so I'm not too concerned about that.

N.B.: The new `PHPCompatibility\Util\TestVersionException` class uses the new-style file and class docblock documentation as proposed in 734.
For that reason, some additional CS exceptions have been set in the `phpcs.xml.dist` file. Those can and will be removed when the big PR to fix 734 for all files is pulled.