Sniff::isClassProperty: bug fix
When a class would be nested in parenthesis, the method did not recognize class properties properly.

To check for this without code duplication, the return of the `Sniff::validDirectScope()` method has changed. It used to always return a boolean. Now it will return `false` if not in a valid direct scope and the `$stackPtr` to the scope if it is.

Includes unit test.