Code Coverage: allow recording for indirect testing of utility methods
A few utility methods indirectly test underlying utility methods too. Similarly,  number of utility methods have been abstracted out of sniffs and are tested via that sniff.

So in select cases, it is valid to record code coverage for the underlying utility method too.
I've added `@covers` statements in a few places to that effect.