Travis: use PHPUnit 7.x for PHP >= 7.2
PHPUnit 8 adds `void` return type declarations to the PHPUnit methods used by PHPCS, making it neigh impossible to make the unit test suite cross-version compatible for the PHP versions supported by PHPCompatibility.

As of the recent update to the Travis images, the images for PHP 7.2+ now ship with PHPUnit 8 which causes the unit tests to fail.

Officially, PHPUnit 7 is compatible with PHP 7.1, 7.2 and 7.3.
Ref: https://phpunit.de/supported-versions.html

So, for Travis images which come natively with PHPUnit 8 (PHP >= 7.2), PHPUnit 7 needs to be installed via Composer.