Travis: test code coverage on the highest supported PHP version
As PHP 7.3 has been released, the build to test code coverage should be run on PHP 7.3.

This should raise the code coverage as for, for instance, the `FlexibleHeredocNowdoc` sniff, code coverage for a significant part of the sniff will only show up when run on PHP 7.3.

In contrast to when the PHP 7.3 build was added, there does appear to be a code coverage driver available for PHP 7.3 by now, so we should be safe to make this change now.

Ref: 764