Travis: remove testing against nightly / PHP 8
There is no PHPUnit version which is currently compatible with PHP 8.
As that either means that the builds for nightly would always fail or - if the unit tests would be skipped -, the only check executed on nightly would be linting the files, I've elected to remove build testing against nightly for the time being.

For more details about PHPUnit vs PHPCS vs PHP 8, see squizlabs/PHP_CodeSniffer 2416