:sparkles: PHP 7.4: new NewTypedProperties sniff
PHP 7.4 will be introducing support for runtime-enforced first-class property type declarations.

Ref: https://wiki.php.net/rfc/typed_properties_v2

This PR introduces a new sniff to detect the use of these and report an error if PHP < 7.4 is intended to still be supported.

The PHP 7.4 implementation does not allow for properties with the `void` or `callable` type.
If these are detected and PHP 7.4 is supposed to be supported, an error will be thrown as well.

Related:
* https://github.com/squizlabs/PHP_CodeSniffer/issues/2413