NewLanguageConstructs: improve T_ELLIPSIS error message text
The `T_ELLIPSIS` or spread operator can be used for both variadic functions as well as for argument unpacking. Support for both was added in PHP 5.6.

In PHP 7.4, a third use case is added: unpacking arrays at array declaration.

This minor change is intended to make the language of the error message clearer so it will be applicable to all use-cases.