RemovedTernaryAssociativity: fix false positive unit test for PHPCS 2.x
While examining something else, I noticed that the `testNoFalsePositives()` unit test for the `RemovedTernaryAssociativity` sniff was being marked as "Risky" by PHPUnit when run in combination with PHPCS 2.x.

Turns out that with how we are running the unit tests in PHPCompatibility for PHPCS 2.x, we don't have access to the token array from within the unit tests, which effectively means that the logic which created the data range to test in the `testNoFalsePositives()` test would yield an empty range of lines, resulting in no assertions being made.

While not as shiny as having the test case file line count calculated automatically, setting it via a property - which will need to be updated whenever the unit test case file is adjusted ! -, fixes this and will allow this unit test to run properly in combination with PHPCS 2.x.