LowPHPCS: add unit test
The `LowPHPCS` sniff is/was the only one which didn't have unit tests. I've now added these.

To prevent code duplication and maintainability issues, I've changed the properties containing the "minimum supported" and "minimum recommended" PHPCS versions to constants so these can be used by both the sniff as well as the unit test.

In addition to this, I noticed a bug where the `reportWidth` could be blank in a unit test situation. This has been fixed.
AFAICS, this is not something which would easily have occured in non-unit test situations.

Includes removing a condition which is no longer necessary now the minimum PHPCS version is 2.3.0.