:sparkles: PHP 7.4: New NewPHPOpenTagEOF sniff
> `<?php` at the end of the file (without trailing newline) will now be interpreted as an opening PHP tag.
> Previously it was interpreted either as `<? php` and resulted in a syntax error (with short_open_tag=1) or was interpreted as a literal `<?php` string (with short_open_tag=0).

Refs:
* https://github.com/php/php-src/blob/30de357fa14480468132bbc22a272aeb91789ba8/UPGRADING#L37-L40
* https://github.com/php/php-src/commit/c9acc9018629d046679fcdccabd4a4f556b87499#diff-7748eb3bfdd3bf962553f6f9f2723c45

## Implementation notes:

This sniff will not work on PHP 5.3 in combination with PHPCS < 2.6.0 with `short_open_tag=On` due to a tokenizer bug in PHPCS itself in those older versions.

Related: 835

The sniff needs different code to detect the issue depending on whether or not `short_open_tag`s is enabled on the system running the sniffs.

By default, the Travis images have `short_open_tag=Off`.

To make sure that both situations are unit tested, two of the existing builds will now change the PHP ini config of the Travis image to ensure that the sniff is also tested with `short_open_tag=On`.