ForbiddenParametersWithSameName: bug fix - variables names are case-sensitive
While it most definitely should be considered bad practice to use the same name in a different case for two different parameters for the same function, it is not problematic from a PHP/PHPCompatibility point of view.

Variables in PHP are _case-sensitive_, so should be treated as such for the purposes of this sniff.

See: https://3v4l.org/9QXAL

Looks like this bug has been in the sniff from the very beginning - 110 -, so in a way quite surprising that we've not had any bug reports about it.

Includes unit test.