NewExceptionsFromToString: detect thrown exceptions from docblock
This adds an additional check based on the docblock to attempt to catch exceptions thrown indirectly from this method.

If the docblock indicates that the function may throw an exception, trust it.

The additional check will only be run if search for the `throw` keyword didn't yield any errors.

Fixes 863