Documentation: improve use of `@internal` tags
For the use of `@internal` tags with comments for dev internal information, it used to be recommended to use `{@internal ... }}` (not the double curlies at the end) format to help IDEs parse these correctly.

This format has since been superseded as the parsing of docblocks in IDEs has improved a lot since then.

The current recommendation is to just use a single curly at the end.

Ref: https://github.com/php-fig/fig-standards/blob/master/proposed/phpdoc-tags.md#55-internal