Utility function tests: switch over to the PHPCSUtils UtilityMethodTestCase base class
This should give significantly faster results.

For a test run on my local machine, these are the results:
* Command: `phpunit --group utilityFunctions`
* Run with the previous commit checked-out: 2.57 min
* Run with this commit checked-out: 7.07 _seconds_.

A requirement for the `UtilityMethodTestCase::getTargetToken()` helper method is that the test markers in test case files for utility methods start with `/* test` - to allow to distinguish between test marker comments and comments which are part of the actual test case -.

To this end, the test markers in the remaining utility method test case files have been renamed.