:sparkles: PHP 7.0: New NewNestedStaticAccess sniff
This sniff only covers a small part of the changes in the PHP 7.0 Uniform Variable Syntax RFC, but it's at least another part which is covered.

> ```php
> // support nested ::
> $foo['bar']::$baz
> $foo::$bar::$baz
> $foo->bar()::baz()
> ```
>
> Static property fetches and method calls can now be applied to any dereferencable expression. E.g. `$foo['bar']::$baz`, `$foo::$bar::$baz` and `$foo->bar()::baz()` are all valid now.

Ref: https://wiki.php.net/rfc/uniform_variable_syntax

The sniff implementation as is being pulled now is supported by various tests run on 3v4l:
https://3v4l.org/up9e3
https://3v4l.org/7tZcL
https://3v4l.org/5uuFR
https://3v4l.org/grUNF
https://3v4l.org/mtWn9
https://3v4l.org/PAtnX
https://3v4l.org/qujtp
https://3v4l.org/Rrteo
https://3v4l.org/GQA2P
https://3v4l.org/2XCUZ
https://3v4l.org/d3vcp

Includes unit tests.

The sniff has been run over a number of large PHP projects to see if any false positives could be found, but didn't find any.

Fixes 946