Various sniffs: make function call check more stable
Make the "is this a non-namespaced function call" check more stable by:
* Checking for parenthesis after it and making sure that those parenthesis don't have an owner (like they would when it is a function declaration).
* Checking previous token via "previous non empty" instead of just disregarding whitespace.
* Make the namespace check more stable by checking the previous non-empty token before the namespace separator instead of the previous token and by checking for both `T_STRING` (`NS\call()`) as well as `T_NAMESPACE` (`namespace\call()`).

I've applied the same fix in a number of places where sniffs check for function calls.

This will be further improved upon once the repo switches over to PHPCSUtils which will have a utility function for this and (planned) an abstract sniff class to use.

Includes some unit tests in the `RemovedFunctions` sniff test case file.

Related to 961