Unit tests: add support for PHPUnit 9.x
In PHPUnit 9.x, the use of `expectException()` for PHP native error/warning/notice/deprecation messages has been deprecated and support will be removed in PHPUnit 10.x.

The dedicated `expectError()`, `expectWarning()`, `expectNotice()` and `expectDeprecation()` methods should be used instead and a warning will be thrown if `expectException()` is used for PHP native exceptions.

This adds a helper method for the one place in the PHPCompatibility test suite where we test that a PHP warning is being thrown handling the cross-version compatibility gracefully.

Includes:
* Composer: updating the `require-dev` section.
* `phpunit.xml.dist`: updating the schema annotation.