New HashAlgorithmsTrait, including implementation
Move functionality which is only needed by select sniffs out of the base `Sniff` class to traits.

This addresses the hash algorithm specific functionality and moves both the property as well as the `getHashAlgorithmParameter()` method to a new `HashAlgorithmsTrait`.

In a future iteration, these sniffs will be further refactored.

The trait is tested via the integration tests for the individual sniffs.