Travis: test on nightly
Up to now, the DealerDirect plugin was a blocker for testing the standard against PHP `nightly`/ PHP 8 as it would refuse to install.

With the new version of the DealerDirect plugin having been tagged and PHPCSUtils allowing for it, this issue has been removed, so we can now run the unit tests against PHP 8 to get early warning about cross-version compatibility issues which need fixing.

The build against `nightly` remains in `allow_failures` for now.

Refs:
* https://github.com/Dealerdirect/phpcodesniffer-composer-installer/releases/tag/v0.7.0
* https://github.com/PHPCSStandards/PHPCSUtils/releases/tag/1.0.0-alpha3