Sniff:: remove the `inClassScope()` method
The `Sniff::inClassScope()` method has become redundant.

All uses of the method have by now been switched out for a call to the `Conditions::hasCondition()` in combination with the `BCTokens::ooScopeTokens()` method, the PHPCompatibility native method - as well as its associated unit tests - can be removed.

Commits which removed the uses of this method:
* #1014 - RemovedPredefinedGlobalVariables: bug fix - static property use outside class context
* #1368 - ForbiddenNames: bug fix - throw error for nested function declarations
* #996 - NewParamTypeDeclarations:: bug fix - throw error for nested function declarations
* #1141 - NewClosure: use PHPCSUtils\Utils\Conditions::hasCondition()
* #981 - NewLateStaticBinding: use PHPCSUtils\Utils\Conditions::hasCondition()