NewTypedProperties: use PHPCSUtils Variables::getMemberProperties()
As PHPCS did not support typed properties in the `File::getMemberProperties()` method until PHPCS 3.5.0, this method could previously not be used to detect typed properties.

The PHPCSUtils version of that same method *does* support typed properties, including support all the way back to PHPCS 2.6.0.

This PR switches out the custom code for using the utility method, greatly simplifying the sniff.