NewNullableTypes: use FunctionDeclarations::getProperties()
The return value  of the PHPCS native `File::getMethodProperties()` method contains a `nullable_return_type` index key since PHPCS 3.3.0.

PHPCSUtils backports this functionality back to PHPCS 2.6.0.

This now allows us to remove the sniffing for the `T_RETURN_TYPE` token and the custom token walking for the analysis of a function return type declaration, which greatly simplifies the sniff.