NewReturnTypeDeclarations: use FunctionDeclarations::getProperties()
The return value  of the PHPCS native `File::getMethodProperties()` method contains a `return_type` and a `return_type_token` index key since PHPCS 3.3.0.

PHPCSUtils backports this functionality back to PHPCS 2.6.0, as well as adding a `return_type_end_token` index to the return value.

This now allows us to remove the sniffing for the `T_RETURN_TYPE` token and simplify the sniff a little.