NonStaticMagicMethods: simplify the logic
Until now, the sniff started from a class/interface/trait token and would walk the whole scope to find all the method declarations contained therein.

This changes the entry-point for the sniff to `T_FUNCTION` combined with a quick "is this an OO method ?" check at the start of the sniff, which allows for removing all the token walking logic from the sniff, making the sniff a little faster.

Note: this commit will be easiest to review with the "ignore whitespace changes" option turned on.